backend/requirements.txt
asgiref==3.2.10
boto3==1.17.68
botocore==1.20.68
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.0.8
django-cors-headers==3.7.0
django-environ==0.4.5
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.3.0
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.26.2
whitenoise==5.2.0
